Maximize change, workflow progress
Added workflow icon to execute ocmmand (will need to hide it for commands which are not workflow or workflow is at the end)
Added lots of code to compenzate for multi select workflows, currently all param input is broken.
Changed maximize window for param input, now uses full width and height and different icon for both.